requirements.txt
absl-py==2.3.1
accelerate==1.9.0
adpbulk==0.1.4
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anndata==0.12.6
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
array-api-compat==1.12.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
cachetools==6.2.2
cell-eval==0.5.42
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.3.1
cloudpickle==3.1.2
comm==0.2.3
contourpy==1.3.3
crc32c==2.7.1
cuda-bindings==12.9.4
cuda-core==0.3.2
cuda-pathfinder==1.3.2
cuda-python==12.9.4
cuda-toolkit==12.9.1
cudf-cu12==25.10.0
cugraph-cu12==25.10.1
cuml-cu12==25.10.0
cupy-cuda12x==13.6.0
cuvs-cu12==25.10.0
cycler==0.12.1
dask==2025.9.1
dask-cuda==25.10.0
dask-cudf-cu12==25.10.0
debugpy==1.8.15
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.34.0
distributed==2025.9.1
distributed-ucxx-cu12==0.46.0
docrep==0.3.2
donfig==0.8.1.post1
executing==2.2.0
faiss-cpu==1.12.0
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.13.1
fonttools==4.61.0
formulaic==1.2.0
formulaic-contrasts==1.0.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
geomloss==0.2.6
gitdb==4.0.12
GitPython==3.1.45
google-crc32c==1.7.1
GPUtil==1.4.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.1
idna==3.11
igraph==0.11.9
ImageIO==2.37.2
importlib_metadata==8.7.0
iniconfig==2.1.0
interface-meta==1.3.0
ipykernel==6.30.0
ipympl==0.9.7
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.6
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keopscore==2.3
kiwisolver==1.4.9
lark==1.2.2
lazy_loader==0.4
legacy-api-wrap==1.5
libcudf-cu12==25.10.0
libcugraph-cu12==25.10.1
libcuml-cu12==25.10.0
libcuvs-cu12==25.10.0
libkvikio-cu12==25.10.0
libraft-cu12==25.10.0
librmm-cu12==25.10.0
libucx-cu12==1.19.0
libucxx-cu12==0.46.0
lightning-utilities==0.15.0
llvmlite==0.44.0
locket==1.0.0
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
narwhals==2.0.1
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numba-cuda==0.19.1
numcodecs==0.16.5
numpy==2.2.6
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cccl-cu12==12.9.27
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvcc-cu12==12.9.86
nvidia-cuda-nvrtc-cu12==12.9.86
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.4.1.4
nvidia-curand-cu12==10.3.10.19
nvidia-cusolver-cu12==11.7.5.82
nvidia-cusparse-cu12==12.5.10.65
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.28.9
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.14
opencv-python==4.12.0.88
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.4
partd==1.4.2
pastel==0.2.1
patsy==1.0.2
pdex==0.1.20
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.3.8
pluggy==1.6.0
poethepoet==0.10.0
polars==1.32.0
POT==0.9.5
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.4.1
protobuf==6.31.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pybind11==3.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeseq2==0.5.2
pyg-lib==0.4.0+pt26cu124
Pygments==2.19.2
pykeops==2.3
pylibcudf-cu12==25.10.0
pylibcugraph-cu12==25.10.1
pylibraft-cu12==25.10.0
pynndescent==0.5.13
pynvjitlink-cu12==0.7.0
pynvml==12.0.0
pyparsing==3.2.5
pytest==8.4.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
pytorch-lightning==2.5.2
pytz==2025.2
pyvers==0.1.0
PyYAML==6.0.3
pyzmq==27.0.0
raft-dask-cu12==25.10.0
rapids-dask-dependency==25.10.0
rapids-logger==0.1.19
rapids_singlecell==0.13.4
rdkit==2025.3.3
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rmm-cu12==25.10.0
rpds-py==0.26.0
safetensors==0.5.3
scanpy==1.11.5
scikit-image==0.25.2
scikit-learn==1.7.2
scikit-misc==0.5.2
scipy==1.16.3
seaborn==0.13.2
Send2Trash==1.8.3
sentry-sdk==2.33.2
session-info2==0.2.3
session_info==1.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
statsmodels==0.14.5
stdlib-list==0.11.1
sympy==1.13.1
tblib==3.2.2
tensordict==0.9.1
terminado==0.18.1
texttable==1.7.0
threadpoolctl==3.6.0
tifffile==2025.10.16
tinycss2==1.4.0
tokenizers==0.21.2
tomlkit==0.13.3
toolz==1.1.0
torch==2.6.0+cu124
torch-geometric==2.6.1
torchaudio==2.6.0+cu124
torchcde==0.2.5
torchcfm==1.0.7
torchdiffeq==0.2.5
torchdyn==1.0.6
torchmetrics==1.8.0
torchsde==0.2.6
torchvision==0.21.0+cu124
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.54.0
treelite==4.4.1
triton==3.2.0
types-python-dateutil==2.9.0.20250708
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
ucx-py-cu12==0.43.0
ucxx-cu12==0.46.0
umap-learn==0.5.9.post2
uri-template==1.3.0
urllib3==2.5.0
wandb==0.21.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
wrapt==1.17.2
yarl==1.22.0
zarr==3.1.5
zict==3.0.0
zipp==3.23.0
